Jython installing ...
* Beginnings of scripts, which were supposed to be in bin/ but I
  forgot about the git exclusions.
* Will refactor to run from new directory and then clean up
  post-installation.

When these are complete, the best option will be to download and
install the fullpackage.  This will make a Jython version of pip
available and allow installing any pure python package.  Including
requests and docutils, which will be installed at the end.  The former
provides for a great many options and the latter opens up more of
reST.  Unfortunately reportlab won't install directly as it needs PIL,
which needs CPython (with the emphasis on C).  Still, Jython can
happily pass things off to a system CPython and that will have to do.

FYI:  ReportLab can be used as a work-around for the FOP 1.1 bug which
results in memory errors preventing PDF generation on some platforms
(e.g. OS X and some Linux distributions, probably some of the BSDs as
well).  This might seem odd to some people until you realise that Java
is not (and never has been) platform independent; that was a goal that
it never achieved, just blurred the line a bit.